requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-cloud-texttospeech==2.16.0
google-generativeai==0.3.2
python-dotenv==1.0.1